requirements.txt
asgiref==3.2.3
aio-pika==6.6.0
Django==2.2.10
djangorestframework==3.11.0
channels==2.4.0
channels-redis==2.4.2
pytz==2019.3
sqlparse==0.3.0
psycopg2==2.8
python-dateutil==2.8.1
django-extensions==2.2.8